Add extensive comments on the safegcd algorithm and implementation
This adds a long comment explaining the algorithm and implementation choices by building
it up step by step in Python. The comment is only present in modinv64_impl.h, but a
reference to it is added in modinv32_impl.h.

Comments in the code are also reworked/added, with references to the long explanation.